pyproject.toml
[project]
name = "misinformation-detection-api"
version = "0.1.0"
description = "API for an LLM trained to detect misinformation in Tweets."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.116.0",
    "passlib[argon2]>=1.7.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy-stubs>=0.4",
    "sqlalchemy[asyncio]>=2.0.41",
    "transformers[torch]>=4.54.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "ipython>=8.37.0",
    "pytest>=8.4.1",
]
